pyproject.toml
[project]
name = "rag-tutor"
version = "0.1.0"
description = "AI-powered math tutor with RAG capabilities"
requires-python = ">=3.10,<3.13"  # langchain-community requires >=3.10, limited to <3.13 to avoid numpy 2.x
# readme is optional - will use if available

[project.optional-dependencies]
# Зависимости для приложения (backend + frontend)
app = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic>=2.7.4,<3.0.0",
    "streamlit>=1.39.0",
    "requests>=2.32.5,<3.0.0",  # langchain-community requires >=2.32.5
    "Pillow==10.1.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "openai>=1.6.1,<2.0.0",
    "langsmith>=0.1.0",
    "e2b-code-interpreter>=1.0.0,<2.0.0",
    "faiss-cpu==1.7.4",
    "sentence-transformers==5.1.2",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
    "gdown==4.7.1",
    "numpy<2.0.0",  # faiss-cpu 1.7.4 requires numpy < 2.0
]

# Зависимости для скриптов обработки данных
scripts = [
    "chromadb==1.3.5",
    "langchain-community==0.4.1",
    "langchain-core==1.1.0",
    "langchain-text-splitters==1.0.0",
    "pdf2image==1.17.0",
    "pytesseract==0.3.13",
    "sentence-transformers==5.1.2",
    "tqdm==4.67.1",
    "numpy<2.0.0",  # sentence-transformers and other packages require numpy < 2.0
]

# Все зависимости вместе
all = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic>=2.7.4,<3.0.0",
    "streamlit>=1.39.0",
    "requests>=2.32.5,<3.0.0",  # langchain-community requires >=2.32.5
    "Pillow==10.1.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community==0.4.1",
    "langchain-text-splitters==1.0.0",
    "openai>=1.6.1,<2.0.0",
    "langsmith>=0.1.0",
    "e2b-code-interpreter>=1.0.0,<2.0.0",
    "faiss-cpu==1.7.4",
    "sentence-transformers==5.1.2",
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
    "gdown==4.7.1",
    "chromadb==1.3.5",
    "pdf2image==1.17.0",
    "pytesseract==0.3.13",
    "tqdm==4.67.1",
    "numpy<2.0.0",  # faiss-cpu 1.7.4 requires numpy < 2.0
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Include app directory and its subdirectories as packages
packages = ["app"]
# Ensure all Python files are included
include = [
    "app/**/*.py",
]

